SunoSync/requirements.txt
requests>=2.31.0
mutagen>=1.47.0
Pillow>=10.0.0
pyperclip>=1.8.2
colorama>=0.4.6
python-vlc>=3.0.20
pyinstaller>=6.0.0
customtkinter>=5.2.0
appdirs>=1.4.4

pyngrok>=7.0.0
